requirements.txt
asgiref==3.6.0
autopep8==2.0.1
chardet==5.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.5
docx2txt==0.8
filelock==3.8.0
fire==0.5.0
Flask==2.2.2
fonttools==4.38.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
pdf2docx==0.5.6
pdf2image==1.16.3
pdfminer==20191125
Pillow==9.4.0
platformdirs==2.5.2
pycodestyle==2.10.0
pycryptodome==3.17
PyDocX==0.9.10
PyMuPDF==1.21.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-docx==0.8.11
six==1.16.0
sqlparse==0.4.3
termcolor==2.2.0
tesseract==0.1.3
tomli==2.0.1
tzdata==2022.7
virtualenv==20.16.3
virtualenvwrapper-win==1.2.7
Werkzeug==2.2.2
